requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
schedule>=1.1.0
pyttsx3>=2.90
SpeechRecognition>=3.8.1
gTTS>=2.3.1
pvporcupine>=2.1.0  # For wake word detection
pyaudio>=0.2.11
python-telegram-bot>=13.7
